requirements.txt
fastapi
uvicorn[standard]
pydantic
opencv-python
numpy
Pillow
scikit-learn
matplotlib
python-multipart
requests
pydantic-settings
pytesseract